added missing deps and get tests to work
